Rename the Webleed module to Heartbleed and move it to the pulse subdomain
Keeping the original name just for images to keep their URLs.
