+++ app/src/Presentation/Heartbleed/Homepage/HomepagePresenter.php
@@ -1,0 +1,16 @@
+<?php
+declare(strict_types = 1);
+
+namespace MichalSpacekCz\Presentation\Heartbleed\Homepage;
+
+use MichalSpacekCz\Presentation\Www\BasePresenter;
+
+final class HomepagePresenter extends BasePresenter
+{
+
+	public function actionDefault(): void
+	{
+		$this->redirectPermanent(':Pulse:Heartbleed:');
+	}
+
+}
