+++ app/src/Presentation/Pulse/Heartbleed/default.latte
@@ -5,7 +5,7 @@
 	<meta charset="utf-8">
 	<meta name="viewport" content="width=device-width, initial-scale=1">
 	<title>We Bleed: Heartbleed Bug in the Czech Republic 2014–2015</title>
-	{styleSheet screen-webleed}
+	{styleSheet screen-heartbleed}
 	<link rel="icon" type="image/png" href="{='/i/icons/webleed.png'|staticUrl}">
 </head>
 <body>
