+++ app/src/Presentation/Pulse/Homepage/default.latte
@@ -2,6 +2,6 @@
 <ul>
 	<li><a n:href="Passwords:storages">Passwords storage disclosures</a></li>
 	<li><a href="https://docs.google.com/spreadsheets/d/1LI1Pk0IwAvD9FE4ShHIU8ajT_NvEtGxO0VFW4OR78TY/edit" title="{_messages.label.banksczhttps}">{_messages.label.bankshttps}</a></li>
-	<li><a n:href=":Webleed:Homepage:">Heartbleed bug in the Czech Republic 2014–2015</a></li>
+	<li><a n:href="Heartbleed:">Heartbleed bug in the Czech Republic 2014–2015</a></li>
 </ul>
 {/define}
